+++ editor/editor-plugin-ufo/src/ufoPlugin.tsx
@@ -1,20 +1,28 @@
-// import { abortUFOMeasurementOnFirstUserInteraction } from './pm-plugins/abortUFOMeasurementOnFirstUserInteraction';
+import { expValEquals } from '@atlaskit/tmp-editor-statsig/exp-val-equals';
+
+import { traceUFOInteractionOnFirstInteraction } from './pm-plugins/traceUFOInteractionOnFirstInteraction';
 import type { UfoPlugin } from './ufoPluginType';
 
 export const ufoPlugin: UfoPlugin = () => ({
 	name: 'ufo',
 
 	pmPlugins() {
-		return [];
+		const isSSR = Boolean(process?.env?.REACT_SSR);
 
-		// This experimental plugin has been abandoned as per ED-27989
-		/*
+		if (isSSR) {
+			// This plugin is not needed in SSR environments - and uses apis which are not available in SSR internally
+			return [];
+		}
+
+		if (expValEquals('cc_editor_interactions_trigger_traceufointeraction', 'cohort', 'control')) {
+			return [];
+		}
+
 		return [
 			{
-				name: 'abortUFOMeasurementOnFirstUserInteraction',
-				plugin: abortUFOMeasurementOnFirstUserInteraction,
+				name: 'traceUFOInteractionOnFirstInteraction',
+				plugin: traceUFOInteractionOnFirstInteraction,
 			},
 		];
-		*/
 	},
 });