+++ linking-platform/smart-card/src/view/FlexibleCard/components/elements/common/base-avatar-group-element/index.tsx
@@ -8,6 +8,7 @@ import { css, cssMap, jsx } from '@compiled/react';
 import { type IntlShape, useIntl } from 'react-intl-next';
 
 import AtlaskitAvatarGroup from '@atlaskit/avatar-group';
+import { fg } from '@atlaskit/platform-feature-flags';
 import { token } from '@atlaskit/tokens';
 
 import { ElementName, SmartLinkSize } from '../../../../../../constants';
@@ -171,6 +172,7 @@ const BaseAvatarGroup = ({
 		<span
 			// eslint-disable-next-line @atlaskit/design-system/consistent-css-prop-usage -- Ignored via go/DSP-18766
 			css={[stylesMap[size], styles]}
+			{...(fg('platform-linking-enable-avatar-data-separator') ? { ['data-separator']: true } : {})}
 			data-smart-element={name}
 			data-smart-element-avatar-group
 			data-testid={testId}